+++ tests/data/preformatted/401_create_row_access_policy.sql
@@ -1,0 +1,4 @@
+create or replace row access policy foo
+on foo.bar.baz
+grant to ('user1', 'user2')
+filter using ( foo = 'bar' )